Final changes before merging dev_corrections into main
