+++ requirements.txt
@@ -1,0 +1,5 @@
+numpy>=1.26.0
+scipy>=1.12.0
+astropy>=6.0.0
+mpdaf>=3.7
+matplotlib>=3.9.0
